fix: estates within districts
