+++ shared/map/index.js
@@ -1,2 +1,3 @@
 export * from './Bounds'
+export * from './Tile'
 export * from './Viewport'
